services/file-process-function/requirements.txt
functions-framework==3.*
google-cloud-monitoring==2.*
cloudevents==1.*
protobuf>=4.0
